ODP-4951 | NIFI-14277 Added credential scope in GCP PubSub and BigQuery processors
